From: Stu Grossman <grossman@juniper.net> To: bochs-developers@lists.sourceforge.net Subject: [Bochs-developers] Fix for configure.in bug
This patches fixes a bug in configure.in which prevents configure from setting
default values for CFLAGS and CXXFLAGS.  Normally, the way autoconf builds
configure, it will set default values for CFLAGS and CXXFLAGS of `-g -O2'.
But, if the user specified them (via the environment), it won't touch them.

The clause in configure.in that tries to set platform-specific flags breaks
this because it always assigns a value (usually blank for most platforms),
which fools the defaulting mechanism later on.


git-svn-id: http://svn.code.sf.net/p/bochs/code@3270 6c8e4198-a580-4521-a02f-c9ae86df1db9