Changed push_32() and push_64() to issue SS(0) for a push outside of stack limits. This is needed for EROS, and probably for L4, as both rely on this SS fault (and the corresponding GP fault) to trigger the switch from small address spaces to large address spaces. The push_16() code was already correct, and I find the inconsistency a bit odd.
I'm not 100% sure about the push_64() change, so I made the change
with a comment but left it a BX_PANIC() rather than switching it to
BX_INFO. I'll ask Peter momentarily to have a look and let me know.

While I was added, changed the push_16() BX_INFO message to be
consistent with the others -- all now say 'push outside stack limits'.


git-svn-id: http://svn.code.sf.net/p/bochs/code@3912 6c8e4198-a580-4521-a02f-c9ae86df1db9