Applied the usb_xhci part of the USB patch by Ben Lunt. Notes by Ben: This emulates a NEC/Renesas uPD720202 2-port (2 socket, 4-port register sets) Root Hub xHCI Host Controller. Many, many thanks to Renesas for their work and effort in helping my research.
I have tested in with my own tests and WinXP Home Edition SP3.

Use port1 and port2 to emulate a Super-speed device, and use port3 and port4
to emulate Low-, Full-, or High-speed devices.

NOTE: The Windows driver for this device uses a proprietary detection
mechanism and we don't have the permission to implement it. That's why it only
works fine with non-Windows guests.


git-svn-id: http://svn.code.sf.net/p/bochs/code@12493 6c8e4198-a580-4521-a02f-c9ae86df1db9